refactor(conf): split code, generate typings
