--- packages/conf/src/fixtures/fallback-locales/.linguirc
@@ -1,4 +1,0 @@
-{
-  "locales": ["en-US"],
-  "fallbackLocale": "en"
-}
