+++ packages/conf/src/index.test.ts
@@ -1,31 +1,25 @@
 import path from "path"
-import { validate } from "jest-validate"
-import {
-  getConfig,
-  replaceRootDir,
-  catalogMigration,
-  configValidation,
-} from "@lingui/conf"
+import { getConfig } from "./getConfig"
+import { makeConfig } from "./makeConfig"
 import { mockConsole, getConsoleMockCalls } from "@lingui/jest-mocks"
+import { LinguiConfig } from "./types"
+import { DeprecatedFallbackLanguage } from "./migrations/fallbackLanguageMigration"
 
-describe("@lingui/conf", function () {
-  it("should return default config", function () {
-    expect.assertions(2)
-
+describe("@lingui/conf", () => {
+  it("should return default config", () => {
     mockConsole((console) => {
       const config = getConfig({
         cwd: path.resolve(__dirname, path.join("fixtures", "valid")),
       })
+      expect(console.error).not.toBeCalled()
       expect(console.warn).not.toBeCalled()
       expect(config).toMatchSnapshot()
     })
   })
 
-  it("should validate `locale`", function () {
-    expect.assertions(2)
-
+  it("should validate `locale`", () => {
     mockConsole((console) => {
-      getConfig()
+      makeConfig({})
       expect(console.warn).not.toBeCalled()
       expect(console.error).toBeCalledWith(
         expect.stringContaining("No locales defined")
@@ -33,113 +27,22 @@ describe("@lingui/conf", function () {
     })
   })
 
-  it("should replace <rootDir>", function () {
-    const config = replaceRootDir(
-      // @ts-ignore
-      {
-        compileNamespace: "cjs",
-        catalogs: [
-          {
-            path: "/",
-            include: ["<rootDir>/src"],
-            exclude: ["<rootDir>/ignored"],
-          },
-        ],
-      },
-      "/Root"
-    )
-
-    expect(config.compileNamespace).toEqual("cjs")
-    expect(config.catalogs).toEqual([
-      {
-        path: "/",
-        include: ["/Root/src"],
-        exclude: ["/Root/ignored"],
-      },
-    ])
-  })
-
-  describe("catalogMigration", function () {
-    it("should show deprecation warning", function () {
-      mockConsole((console) => {
-        validate(
-          {
-            localeDir: "./locale",
-            srcPathDirs: ["./src"],
-            srcPathIgnorePatterns: ["/node_modules/"],
-          },
-          configValidation
-        )
-        expect(getConsoleMockCalls(console.warn)).toMatchSnapshot()
+  describe("config file", () => {
+    it("searches for a .linguirc config file", () => {
+      const config = getConfig({
+        cwd: path.resolve(__dirname, path.join("fixtures", "valid")),
       })
+      expect(config.locales).toEqual(["en-gb"])
     })
 
-    it("shouldn't provide default config if no obsolete config is defined", function () {
-      const config = {}
-
-      // @ts-ignore
-      expect(catalogMigration(config)).toEqual({})
-    })
-
-    it("should normalize string localeDir", function () {
-      const config = {
-        localeDir: "./locales",
-      }
-
-      // @ts-ignore
-      expect(catalogMigration(config)).toEqual({
-        catalogs: [
-          {
-            path: "locales/{locale}/messages",
-            include: ["<rootDir>"],
-            exclude: ["**/node_modules/**"],
-          },
-        ],
+    it("allows specific config file to be loaded with configPath parameter", () => {
+      const config = getConfig({
+        configPath: path.resolve(
+          __dirname,
+          path.join("fixtures", "valid", "custom.config.js")
+        ),
       })
-    })
-
-    it("should migrate srcPathDirs and srcPathIgnorePatterns", function () {
-      const config = {
-        localeDir: "locales",
-        srcPathDirs: ["src"],
-        srcPathIgnorePatterns: ["src/node_modules/"],
-      }
-
-      // @ts-ignore
-      expect(catalogMigration(config)).toEqual({
-        catalogs: [
-          {
-            path: "locales/{locale}/messages",
-            include: ["src"],
-            exclude: ["src/node_modules/"],
-          },
-        ],
-      })
-    })
-  })
-
-  describe("config file", function () {
-    it("searches for a .linguirc config file", function () {
-      // hide validation warning about missing locales
-      mockConsole(() => {
-        const config = getConfig({
-          cwd: path.resolve(__dirname, path.join("fixtures", "valid")),
-        })
-        expect(config.locales).toEqual(["en-gb"])
-      })
-    })
-
-    it("allows specific config file to be loaded with configPath parameter", function () {
-      // hide validation warning about missing locales
-      mockConsole(() => {
-        const config = getConfig({
-          configPath: path.resolve(
-            __dirname,
-            path.join("fixtures", "valid", "custom.config.js")
-          ),
-        })
-        expect(config.locales).toEqual(["cs", "sk"])
-      })
+      expect(config.locales).toEqual(["cs", "sk"])
     })
 
     it("loads TypeScript config", () => {
@@ -152,31 +55,54 @@ describe("@lingui/conf", function () {
       expect(config.locales).toEqual(["pl"])
     })
 
-    describe("fallbackLocales logic", () => {
-      it("if fallbackLocale is defined, we use the default one on fallbackLocales", () => {
+    describe("fallbackLocale deprecation", () => {
+      it("if deprecated fallbackLocale is defined, we set fallbackLocales.default and show message", () => {
         mockConsole((console) => {
-          const config = getConfig({
-            configPath: path.resolve(
-              __dirname,
-              path.join("fixtures", "fallback-locales", ".linguirc")
-            ),
-          })
+          const config = makeConfig({
+            locales: ["en-US"],
+            fallbackLocale: "en",
+          } as LinguiConfig & DeprecatedFallbackLanguage)
+
           expect(config.fallbackLocales.default).toEqual("en")
           expect(getConsoleMockCalls(console.warn)).toMatchSnapshot()
         })
       })
+    })
 
-      it("if fallbackLocales default is defined, we dont build the cldr", () => {
-        const config = getConfig({
-          configPath: path.resolve(
-            __dirname,
-            path.join("fixtures", "valid", ".fallbacklocalesrc")
-          ),
+    describe("Build parent cldr fallbackLocales", () => {
+      it("if fallbackLocales.default is defined, we dont build the cldr", () => {
+        const config = makeConfig({
+          locales: ["en-US", "es-MX"],
+          fallbackLocales: {
+            "en-US": ["en"],
+            default: "en",
+          },
         })
+
         expect(config.fallbackLocales).toEqual({
           "en-US": ["en"],
           default: "en",
         })
+      })
+
+      it("if fallbackLocales.default is not defined, we build the cldr", () => {
+        const config = makeConfig({
+          locales: ["en-US", "es-MX"],
+        })
+
+        expect(config.fallbackLocales).toEqual({
+          "en-US": "en",
+          "es-MX": "es",
+        })
+      })
+
+      it("if fallbackLocales = false, we dont build the cldr", () => {
+        const config = makeConfig({
+          locales: ["en-US", "es-MX"],
+          fallbackLocales: false,
+        })
+
+        expect(config.fallbackLocales).toEqual({})
       })
     })
   })
