+++ tsconfig.json
@@ -16,7 +16,7 @@
       "@lingui/core/compile": ["./packages/core/src/compile/index.ts"],
       "@lingui/cli/api": ["./packages/cli/src/api"],
       "@lingui/react": ["./packages/react/src"],
-      "@lingui/conf": ["./packages/conf"],
+      "@lingui/conf": ["./packages/conf/src"],
       "@lingui/macro": ["./packages/macro/src"]
     }
   },
